Removes TopComponent, organises Form structure
