+++ src/components/DataView/DataViewTypes.ts
@@ -5,7 +5,7 @@ import { MenuItemProps } from "../MenuItem";
 import * as React from "react";
 import { DataViewActionsButtonRowProps } from "./DataViewActionsButtonRow";
 import { DataViewBulkActionsButtonsRowProps } from "./DataViewBulkActionsButtonsRow";
-import { TitleWrapperProps } from "@root/forms/TopComponent/Utils/TitleWrapperTypes";
+import { TitleWrapperProps } from "@root/components/Title";
 
 export interface DataViewColumnTransformArgs<T = unknown> {
 	/** The value of the specific column that is being transformed */
