+++ src/components/Form/Col/ColTypes.ts
@@ -1,0 +1,13 @@
+import { FieldDef } from "../FormTypes";
+
+export interface ColPropsTypes {
+	col: (string | FieldDef)[];
+	// TODO Use something other than any
+	state: any;
+	fieldsDef: FieldDef[];
+	dispatch: any;
+	colsInRow?: number;
+	colIdx?: number;
+	rowIdx?: number;
+	sectionIdx?: number;
+}
