+++ src/components/Form/Col/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./Col"
+export * from "./ColTypes"
