+++ src/components/Form/Layout/LayoutStyles.tsx
@@ -1,0 +1,9 @@
+import styled from "styled-components";
+
+export const StyledLayout = styled.div`
+	margin: 0px;
+	max-width: 1160px;
+	display: grid;
+	grid-template-columns: repeat(1,minmax(0,1fr));
+	gap: 24px 40px;
+`;
