+++ src/components/Form/Layout/LayoutTypes.ts
@@ -1,0 +1,9 @@
+import { FieldDef, SectionDef } from "../FormTypes";
+
+export interface LayoutProps {
+    state: any;
+    dispatch: any;
+    fields: FieldDef[];
+    sections: SectionDef[];
+    registerRef?: (ref: HTMLElement) => () => void
+  }
