+++ src/components/Form/Row/RowTypes.ts
@@ -1,0 +1,10 @@
+import { FieldDef } from "../FormTypes";
+
+export interface RowPropTypes {
+	row: string[][];
+	state: any;
+	fieldsDef: FieldDef[];
+	dispatch: any;
+	rowIdx?: number;
+	sectionIdx?: number;
+}
