+++ src/components/Form/Section/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./Section";
+export * from "./SectionTypes"
