+++ src/components/Form/Top/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./Top";
+export * from "./TopTypes"
