+++ src/components/SummaryPageTopComponent/SummaryPageTopComponent.tsx
@@ -16,7 +16,7 @@ import StarBorder from "@mui/icons-material/StarBorderRounded";
 // Components
 import Image from "@root/components/Image";
 import Button from "../Button";
-import TitleWrapper from "@root/forms/TopComponent/Utils/TitleWrapper";
+import TitleWrapper from "@root/components/Title";
 import evaluateShow from "@root/utils/show/evaluateShow";
 import ButtonRow from "../ButtonRow/ButtonRow";
