+++ src/components/Title/index.ts
@@ -1,0 +1,3 @@
+export { default } from "./TitleWrapper";
+export { default as TitleBackButton } from "./TitleBackButton"
+export * from "./TitleWrapperTypes";
