--- src/forms/TopComponent/index.ts
@@ -1,3 +1,0 @@
-export { default } from "./TopComponent";
-export * from "./TopComponent";
-export * from "./TopComponentTypes";
